+++ nengo_dl/simulator.py
@@ -937,7 +937,7 @@ class Simulator(object):
 
         logger.info("Gradient check passed")
 
-    def trange(self, dt=None):
+    def trange(self, sample_every=None, dt=None):
         """
         Create a vector of times matching probed data.
 
@@ -946,14 +946,23 @@ class Simulator(object):
 
         Parameters
         ----------
-        dt : float, optional
-            The sampling period of the probe to create a range for;
-            if None, the simulator's ``dt`` will be used.
+        sample_every : float, optional (Default: None)
+            The sampling period of the probe to create a range for.
+            If None, a time value for every ``dt`` will be produced.
         """
 
-        dt = self.dt if dt is None else dt
-        n_steps = int(self.n_steps * (self.dt / dt))
-        return dt * np.arange(1, n_steps + 1)
+        # TODO: can remove this if we upgrade minimum nengo version
+        if dt is not None:
+            if sample_every is not None:
+                raise ValidationError(
+                    "Cannot specify both `dt` and `sample_every`. "
+                    "Use `sample_every` only.", attr="dt", obj=self)
+            warnings.warn("`dt` is deprecated. Use `sample_every` instead.")
+            sample_every = dt
+
+        period = 1 if sample_every is None else sample_every / self.dt
+        steps = np.arange(1, self.n_steps + 1)
+        return self.dt * steps[steps % period < 1]
 
     def close(self):
         """